refactor: move AppMode class to config_api/config_models.py
Move AppMode enum from openhands/server/types.py (legacy V0) to
openhands/app_server/config_api/config_models.py (V1 app server).

Updated all 31 import paths across:
- openhands/app_server/
- openhands/server/config/
- enterprise/
- tests/unit/
