+++ enterprise/server/auth/gitlab_sync.py
@@ -5,7 +5,7 @@ from sqlalchemy import select
 
 from openhands.app_server.integrations.service_types import ProviderType
 from openhands.app_server.utils.logger import openhands_logger as logger
-from openhands.server.types import AppMode
+from openhands.app_server.config_api.config_models import AppMode
 
 
 async def _user_has_gitlab_provider(user_id: str) -> bool:
