+++ enterprise/server/config.py
@@ -24,7 +24,7 @@ from server.constants import DEPLOYMENT_MODE
 
 from openhands.app_server.integrations.service_types import ProviderType
 from openhands.server.config.server_config import ServerConfig
-from openhands.server.types import AppMode
+from openhands.app_server.config_api.config_models import AppMode
 
 
 def sign_token(payload: dict[str, object], jwt_secret: str, algorithm='HS256') -> str:
