+++ openhands/app_server/integrations/bitbucket/service/repos.py
@@ -2,9 +2,9 @@ import re
 from typing import Any
 from urllib.parse import urlparse
 
+from openhands.app_server.config_api.config_models import AppMode
 from openhands.app_server.integrations.bitbucket.service.base import BitBucketMixinBase
 from openhands.app_server.integrations.service_types import Repository, SuggestedTask
-from openhands.server.types import AppMode
 
 
 class BitBucketReposMixin(BitBucketMixinBase):
