+++ openhands/app_server/integrations/bitbucket_data_center/service/repos.py
@@ -1,11 +1,11 @@
 from typing import Any
 from urllib.parse import urlparse
 
+from openhands.app_server.config_api.config_models import AppMode
 from openhands.app_server.integrations.bitbucket_data_center.service.base import (
     BitbucketDCMixinBase,
 )
 from openhands.app_server.integrations.service_types import Repository, SuggestedTask
-from openhands.server.types import AppMode
 
 
 class BitbucketDCReposMixin(BitbucketDCMixinBase):
