+++ openhands/app_server/integrations/forgejo/service/repos.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
+from openhands.app_server.config_api.config_models import AppMode
 from openhands.app_server.integrations.forgejo.service.base import ForgejoMixinBase
 from openhands.app_server.integrations.service_types import Repository
-from openhands.server.types import AppMode
 
 
 class ForgejoReposMixin(ForgejoMixinBase):
