+++ openhands/app_server/integrations/gitlab/service/repos.py
@@ -1,3 +1,4 @@
+from openhands.app_server.config_api.config_models import AppMode
 from openhands.app_server.integrations.gitlab.service.base import GitLabMixinBase
 from openhands.app_server.integrations.service_types import (
     OwnerType,
@@ -5,7 +6,6 @@ from openhands.app_server.integrations.service_types import (
     Repository,
 )
 from openhands.app_server.utils.logger import openhands_logger as logger
-from openhands.server.types import AppMode
 
 
 class GitLabReposMixin(GitLabMixinBase):
