+++ tests/unit/app_server/test_webhook_router_auth.py
@@ -11,6 +11,7 @@ import pytest
 from fastapi import FastAPI, HTTPException, status
 from fastapi.testclient import TestClient
 
+from openhands.app_server.config_api.config_models import AppMode
 from openhands.app_server.event_callback.webhook_router import (
     router as webhook_router,
 )
@@ -23,7 +24,6 @@ from openhands.app_server.user.specifiy_user_context import (
     USER_CONTEXT_ATTR,
     SpecifyUserContext,
 )
-from openhands.server.types import AppMode
 
 
 class MockRequestState:
