+++ tests/unit/integrations/bitbucket/test_bitbucket_repos.py
@@ -5,6 +5,7 @@ from unittest.mock import patch
 import pytest
 from pydantic import SecretStr
 
+from openhands.app_server.config_api.config_models import AppMode
 from openhands.app_server.integrations.bitbucket.bitbucket_service import (
     BitBucketService,
 )
@@ -12,7 +13,6 @@ from openhands.app_server.integrations.service_types import OwnerType, Repositor
 from openhands.app_server.integrations.service_types import (
     ProviderType as ServiceProviderType,
 )
-from openhands.server.types import AppMode
 
 
 @pytest.fixture
