+++ tests/unit/integrations/bitbucket_data_center/test_bitbucket_dc.py
@@ -6,11 +6,11 @@ from unittest.mock import patch
 import pytest
 from pydantic import SecretStr
 
+from openhands.app_server.config_api.config_models import AppMode
 from openhands.app_server.integrations.bitbucket_data_center.bitbucket_dc_service import (
     BitbucketDCService,
 )
 from openhands.app_server.integrations.service_types import AuthenticationError, User
-from openhands.server.types import AppMode
 
 # ── init / BASE_URL ───────────────────────────────────────────────────────────
