+++ tests/unit/integrations/github/test_github_service.py
@@ -4,6 +4,7 @@ import httpx
 import pytest
 from pydantic import SecretStr
 
+from openhands.app_server.config_api.config_models import AppMode
 from openhands.app_server.integrations.github.github_service import GitHubService
 from openhands.app_server.integrations.service_types import (
     AuthenticationError,
@@ -12,7 +13,6 @@ from openhands.app_server.integrations.service_types import (
     Repository,
     User,
 )
-from openhands.server.types import AppMode
 
 
 @pytest.mark.asyncio
