+++ tests/unit/server/routes/test_mcp_routes.py
@@ -3,9 +3,9 @@ from unittest.mock import AsyncMock, patch
 
 import pytest
 
+from openhands.app_server.config_api.config_models import AppMode
 from openhands.app_server.integrations.service_types import GitService
 from openhands.app_server.mcp.mcp_router import get_conversation_link
-from openhands.server.types import AppMode
 
 
 def test_mcp_server_no_stateless_http_deprecation_warning():
