+++ tests/unit/test_forgejo_service.py
@@ -4,6 +4,7 @@ import httpx
 import pytest
 from pydantic import SecretStr
 
+from openhands.app_server.config_api.config_models import AppMode
 from openhands.app_server.integrations.forgejo.forgejo_service import ForgejoService
 from openhands.app_server.integrations.service_types import (
     ProviderType,
@@ -11,7 +12,6 @@ from openhands.app_server.integrations.service_types import (
     RequestMethod,
     User,
 )
-from openhands.server.types import AppMode
 
 
 @pytest.fixture
